+++ components/skus/browser/BUILD.gn
@@ -21,7 +21,7 @@ static_library("browser") {
     "switches.h",
   ]
 
-  public_deps = [ "brave-rewards-cxx/src" ]
+  public_deps = [ "rs/cxx" ]
 
   deps = [
     "//base",
