+++ components/skus/browser/rs/lib/tests/e2e.rs
