+++ components/skus/browser/rs/wasm/.gitignore
@@ -1,2 +1,3 @@
 /target
 **/*.rs.bk
+pkg/skus_sdk*
