+++ components/skus/browser/skus_sdk_context_impl.cc
@@ -10,7 +10,7 @@
 
 #include "base/logging.h"
 #include "base/threading/sequenced_task_runner_handle.h"
-#include "brave/components/skus/browser/brave-rewards-cxx/src/lib.rs.h"
+#include "brave/components/skus/browser/rs/cxx/src/lib.rs.h"
 #include "brave/components/skus/browser/pref_names.h"
 #include "brave/components/skus/browser/skus_sdk_fetcher_impl.h"
 #include "components/pref_registry/pref_registry_syncable.h"
