+++ components/skus/browser/skus_sdk_fetcher_impl.h
@@ -9,7 +9,7 @@
 #include <memory>
 #include <string>
 
-#include "brave/components/skus/browser/brave-rewards-cxx/src/lib.rs.h"
+#include "brave/components/skus/browser/rs/cxx/src/lib.rs.h"
 #include "brave/third_party/rust/cxx/include/cxx.h"
 #include "services/network/public/cpp/resource_request.h"
 #include "services/network/public/cpp/shared_url_loader_factory.h"
