+++ components/skus/browser/skus_sdk_service.cc
@@ -10,8 +10,8 @@
 #include "base/json/json_reader.h"
 #include "base/task/post_task.h"
 #include "base/threading/sequenced_task_runner_handle.h"
-#include "brave/components/skus/browser/brave-rewards-cxx/src/lib.rs.h"
-#include "brave/components/skus/browser/brave-rewards-cxx/src/shim.h"
+#include "brave/components/skus/browser/rs/cxx/src/lib.rs.h"
+#include "brave/components/skus/browser/rs/cxx/src/shim.h"
 #include "brave/components/skus/browser/pref_names.h"
 #include "brave/components/skus/browser/skus_sdk_context_impl.h"
 #include "brave/components/skus/browser/skus_utils.h"
